Add flowWater subroutine
Add a first flowWater implementation, which allows the water to flow
correctly through the pipe, stopping if it doesn't find only one free
path to follow on the playfield.

Timer is not implemented yet, and there are still some bugs on starting
pipes position, mainly resulting in impossible to start configuration.