Refactor banks and first time subroutine
Include a first routine and, since rom space wasn't enough, refactor
some DrawPipe function (that might be on vblank later) to be smaller
enough to share bank 2 with other functions and kernel neede space.

Two functions were moved to bank 2 and the number of banks were changed
to 2, only (8kb). It seems a good target for a first version of the
game, assuming we still have some space to recover if we apply the same
strategy for saving rom space to _DrawMiniPipes.