Improve timer and waterFlow
Add a better timer, since only checking the second variable will not be
good enough for the water flow timing.

Also, add new features for the waterFlow subroutine, it can now detect
double-pipes and level finish. Add a return variable, so the main loop
can detect and do whatever it wants based on the result.