Add locked logic for start pipes
Add supporting function and logic for blocking the use of starting
pipes. Some problems arised when using {} with variables, so avoiding it
on all parts. Using manual | or & instead.

Some extra functions for dealing with different indexes had to be added,
some dims were made to avoid variable confusion. The next step would be
detecting a new pipe was used during water flow, and them marking it as
locked.