Free 830 bytes on rom using subroutines
Replace inlined pfpixel calls on DrawMiniPipe subroutine by two nested
subroutines, one for ON and other for OFF. Small change on line 2 (to
always use arg1 and arg4 on pfpixel), and go, from 41 bytes free to 871
bytes.

Such space is important, specially to avoid bank switch on game logic
and to allow the game to stay as 8kb size.