Split _DrawPipe in parts
Same as before, due performance reasons, split _DrawPipe in 3 parts,
what always seems to have less than 1000 cycles.

New variables, exclusive to this subroutine, were added, to keep the
stat safe (arg5 is the only one not needed), and add the 3 calls on the
same place, for now. Two options could be followed to use it to reduce
the cycles usage and keep under the limit imposed by the hardware:

- Draw using vblank, only block drawing a new one. Only likely to happen
  on a rewrite (what could, and maybe should, be treated).
- Lock the player and the water for 3 frames

The minipipe part or the level start wasn't touched yet.