Add simple sounds effects
Add 4 sounds effects:

- Small warning for water flow start
- Small bip for each flow
- Defeat atari-style explosion
- Very bad victory piano jingle

Timer subroutine was slightly changed: it can now return a different
value for the init time case. It wasn't needed before (it was just
starting to flow at that moment, no difference from the normal timer
return was needed).

A separate loop was added to deal with the final sounds.